requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pymongo==4.6.0
motor==3.3.1
pydantic==2.4.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
requests==2.31.0
